+++ test/properties/wrap-for-optimizing-test.js
@@ -14,7 +14,7 @@ vows.describe(wrapForOptimizing)
             ['property-value', '0'],
             ['property-value', '0']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -64,7 +64,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', 'color'],
             ['property-value', 'red']
           ]
-        ]);
+        ], true);
       },
       'has two wraps': function (wrapped) {
         assert.lengthOf(wrapped, 2);
@@ -82,7 +82,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', 'color'],
             ['property-value', 'red']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -101,7 +101,7 @@ vows.describe(wrapForOptimizing)
             ['property-value', '/'],
             ['property-value', '2px']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -124,7 +124,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', '--color'],
             ['property-value', 'red']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -134,6 +134,40 @@ vows.describe(wrapForOptimizing)
       },
       'is not a block': function (wrapped) {
         assert.isFalse(wrapped[0].block);
+      }
+    },
+    'variable reference': {
+      'topic': function () {
+        return wrapForOptimizing([
+          [
+            'property',
+            ['property-name', 'color'],
+            ['property-value', 'var(--red)']
+          ]
+        ], true);
+      },
+      'has one wrap': function (wrapped) {
+        assert.lengthOf(wrapped, 1);
+      },
+      'has name': function (wrapped) {
+        assert.deepEqual(wrapped[0].name, 'color');
+      },
+      'has value': function (wrapped) {
+        assert.deepEqual(wrapped[0].value, [['property-value', 'var(--red)']]);
+      }
+    },
+    'variable reference when variables are ignored': {
+      'topic': function () {
+        return wrapForOptimizing([
+          [
+            'property',
+            ['property-name', 'color'],
+            ['property-value', 'var(--red)']
+          ]
+        ], false);
+      },
+      'has one wrap': function (wrapped) {
+        assert.lengthOf(wrapped, 0);
       }
     },
     'variable block': {
@@ -158,7 +192,7 @@ vows.describe(wrapForOptimizing)
               ]
             ]
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -183,7 +217,7 @@ vows.describe(wrapForOptimizing)
               ]
             ]
           ]
-        ]);
+        ], true);
       },
       'is a block': function (wrapped) {
         assert.isTrue(wrapped[0].block);
@@ -196,7 +230,7 @@ vows.describe(wrapForOptimizing)
             'property',
             ['property-name', 'margin']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -216,7 +250,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', 'margin'],
             ['property-value', '0!important']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -237,7 +271,7 @@ vows.describe(wrapForOptimizing)
             ['property-value', '0'],
             ['property-value', '!important']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -258,7 +292,7 @@ vows.describe(wrapForOptimizing)
             ['property-value', '0!'],
             ['property-value', 'important']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -280,7 +314,7 @@ vows.describe(wrapForOptimizing)
             ['property-value', '!'],
             ['property-value', 'important']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -303,7 +337,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', '_color'],
             ['property-value', 'red']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -323,7 +357,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', '*color'],
             ['property-value', 'red']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -343,7 +377,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', 'margin'],
             ['property-value', '0\\9']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -363,7 +397,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', 'margin'],
             ['property-value', '0']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -384,7 +418,7 @@ vows.describe(wrapForOptimizing)
             ['property-value', '0'],
             ['property-value', '\\9']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -404,7 +438,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', 'margin'],
             ['property-value', '0!ie']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -427,7 +461,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', 'margin'],
             ['property-value', '0 !ie']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -451,7 +485,7 @@ vows.describe(wrapForOptimizing)
             ['property-value', '0'],
             ['property-value', '!ie']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -474,7 +508,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', 'color'],
             ['property-value', 'red\\9!important']
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);
@@ -497,7 +531,7 @@ vows.describe(wrapForOptimizing)
             ['property-name', 'color', [[1, 2, undefined]]],
             ['property-value', 'red', [[1, 2, undefined]]]
           ]
-        ]);
+        ], true);
       },
       'has one wrap': function (wrapped) {
         assert.lengthOf(wrapped, 1);